implements control exceptions and logger.

+++ UserService/src/main/java/com/ironhack/UserService/service/UserService.java
@@ -3,13 +3,18 @@ package com.ironhack.UserService.service;
 import com.ironhack.UserService.exceptions.DataNotFoundException;
 import com.ironhack.UserService.model.User;
 import com.ironhack.UserService.repository.UserRepository;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
+
+    private static final Logger LOGGER = LogManager.getLogger(UserService.class);
 
     @Autowired
     private UserRepository userRepository;
@@ -19,7 +24,12 @@ public class UserService {
      * @return
      */
     public List<User> findAll(){
-        return userRepository.findAll();
+        try {
+            return userRepository.findAll();
+        } catch (Exception ex) {
+            LOGGER.error(ex);
+            throw ex;
+        }
     }
 
     /**
@@ -28,7 +38,18 @@ public class UserService {
      * @return
      */
     public User findById(String id){
-        return userRepository.findById(id).orElseThrow(() -> new DataNotFoundException("This id Club not found."));
+        try {
+            Optional<User> userOptional = userRepository.findById(id);
+
+            if (userOptional.isPresent()){
+                return userOptional.get();
+            } else {
+                return null;
+            }
+        } catch (Exception ex) {
+            LOGGER.error(ex);
+            throw ex;
+        }
     }
 
     /**
@@ -37,7 +58,12 @@ public class UserService {
      * @return
      */
     public User create(User user) {
-        return userRepository.save(user);
+        try {
+            return userRepository.save(user);
+        } catch (Exception ex) {
+            LOGGER.error(ex);
+            throw ex;
+        }
     }
 
     /**
@@ -46,11 +72,19 @@ public class UserService {
      * @param user
      */
     public void update(String id, User user){
-        User userFound = findById(id);
+        try {
+            Optional<User> userOptional = userRepository.findById(id);
 
-        user.setId(userFound.getId());
+            if (userOptional.isPresent()) {
+                User userFound = userOptional.get();
 
-        userRepository.save(user);
+                user.setId(userFound.getId());
+                userRepository.save(user);
+            }
+        } catch (Exception ex) {
+            LOGGER.error(ex);
+            throw ex;
+        }
     }
 
     /**
@@ -58,8 +92,17 @@ public class UserService {
      * @param id
      */
     public void delete(String id){
-        User userFound = findById(id);
+        try {
+            Optional<User> userOptional = userRepository.findById(id);
 
-        userRepository.delete(userFound);
+            if (userOptional.isPresent()) {
+                User userFound = userOptional.get();
+
+                userRepository.delete(userFound);
+            }
+        } catch (Exception ex) {
+            LOGGER.error(ex);
+            throw ex;
+        }
     }
 }
